test: split PDF visual equation scope safety coverage
